更新商户号

+++ service/src/main/java/com/eghm/service/sys/impl/SysRoleServiceImpl.java
@@ -4,7 +4,9 @@ import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.eghm.common.enums.ErrorCode;
 import com.eghm.common.enums.ref.RoleType;
+import com.eghm.common.exception.BusinessException;
 import com.eghm.mapper.SysOperatorRoleMapper;
 import com.eghm.mapper.SysRoleMapper;
 import com.eghm.model.SysOperatorRole;
@@ -15,6 +17,7 @@ import com.eghm.model.dto.role.RoleQueryRequest;
 import com.eghm.service.sys.SysRoleService;
 import com.eghm.utils.DataUtil;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
@@ -27,6 +30,7 @@ import java.util.stream.Collectors;
  */
 @Service("sysRoleService")
 @AllArgsConstructor
+@Slf4j
 public class SysRoleServiceImpl implements SysRoleService {
 
     private final SysRoleMapper sysRoleMapper;
@@ -48,18 +52,26 @@ public class SysRoleServiceImpl implements SysRoleService {
 
     @Override
     public void update(RoleEditRequest request) {
+        this.redoRole(request.getRoleName(), request.getId());
         SysRole role = DataUtil.copy(request, SysRole.class);
         sysRoleMapper.updateById(role);
     }
 
     @Override
     public void delete(Long id) {
+        SysRole sysRole = sysRoleMapper.selectById(id);
+        if (sysRole.getRoleType() != RoleType.COMMON) {
+            log.info("该角色为系统默认角色,无法删除 [{}] [{}]", id, sysRole.getRoleType());
+            throw new BusinessException(ErrorCode.ROLE_FORBID_DELETE);
+        }
         sysRoleMapper.deleteById(id);
     }
 
     @Override
     public void create(RoleAddRequest request) {
+        this.redoRole(request.getRoleName(), null);
         SysRole role = DataUtil.copy(request, SysRole.class);
+        role.setRoleType(RoleType.COMMON);
         sysRoleMapper.insert(role);
     }
 
@@ -108,5 +120,21 @@ public class SysRoleServiceImpl implements SysRoleService {
     @Override
     public boolean isAdminRole(Long operatorId) {
         return sysRoleMapper.countByRoleType(operatorId, RoleType.ADMINISTRATOR.getValue()) > 0;
+    }
+
+    /**
+     * 校验角色是否重复
+     * @param name 角色名称
+     * @param id id 编辑时不能为空
+     */
+    public void redoRole(String name, Long id) {
+        LambdaQueryWrapper<SysRole> wrapper = Wrappers.lambdaQuery();
+        wrapper.eq(SysRole::getRoleName, name);
+        wrapper.ne(id != null, SysRole::getId, id);
+        Integer count = sysRoleMapper.selectCount(wrapper);
+        if (count > 0) {
+            log.error("角色名称重复[{}] [{}]", name, id);
+            throw new BusinessException(ErrorCode.ROLE_NAME_REDO);
+        }
     }
 }
